% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_integer.R
\name{factor_integer}
\alias{factor_integer}
\title{Provides factors of integer}
\usage{
factor_integer(n)
}
\arguments{
\item{n}{number to be factored}
}
\value{
a numeric vector of factors
}
\description{
Provides factors of integer
}
\examples{
factor_integer(6)
factor_integer(7)
factor_integer(24)
}
