% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fibonaccis.R
\name{fibonaccis}
\alias{fibonaccis}
\title{Compute the first n Fibonacci numbers}
\usage{
fibonaccis(n)
}
\arguments{
\item{n}{The number of Fibonacci numbers to compute}
}
\value{
A numeric vector of length n
}
\description{
Compute the first n Fibonacci numbers
}
\examples{
fibonaccis(5)
fibonaccis(8)
fibonaccis(0)
}
