% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_factors.R
\name{get_factors}
\alias{get_factors}
\title{Provides factors of integer}
\usage{
get_factors(n)
}
\arguments{
\item{n}{number to be factored}
}
\value{
a numeric vector of factors
}
\description{
Provides factors of integer
}
\examples{
get_factors(6)
get_factors(7)
get_factors(24)
}
