% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fibonaccis.R
\name{get_first_n_fibonaccis}
\alias{get_first_n_fibonaccis}
\title{Provides first n fibonacci numbers}
\usage{
get_first_n_fibonaccis(n, firstTwo = c(1L, 1L))
}
\arguments{
\item{n}{the number of fibonacci numbers to return}

\item{firstTwo}{the first two number in the sequence (default is \code{c(1, 1)})}
}
\value{
a numeric of length n containing the sequence of fibonacci numbers
}
\description{
Provides first n fibonacci numbers
}
\examples{
get_first_n_fibonaccis(1)
get_first_n_fibonaccis(2)
get_first_n_fibonaccis(5)

# Lucas numbers
get_first_n_fibonaccis(5, firstTwo = c(1, 3))

}
\seealso{
\code{\link[=get_nth_fibonacci]{get_nth_fibonacci()}}
}
