% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_primes.R
\name{get_first_n_primes}
\alias{get_first_n_primes}
\title{Provides the first n prime numbers}
\usage{
get_first_n_primes(n)
}
\arguments{
\item{n}{the number of primes to return}
}
\value{
a numeric vector of length n containing the first n primes
}
\description{
uses the prime number theorem to determine the
}
\examples{
get_first_n_primes(1)
get_first_n_primes(3)
get_first_n_primes(10)
}
