% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fibonaccis.R
\name{get_nth_fibonacci}
\alias{get_nth_fibonacci}
\title{Provides nth fibonnacci number}
\usage{
get_nth_fibonacci(n)
}
\arguments{
\item{n}{the fibonnaci number returned}
}
\value{
a single numeric
}
\description{
Uses explicit closed-form formula to compute nth fibonacci number.
}
\section{May be useful for}{

\itemize{
\item Problem 25: 1000-digit Fibonacci number
}
}

\examples{
get_nth_fibonacci(5)
get_nth_fibonacci(500)
all(get_first_n_fibonaccis(5) == sapply(1:5, get_nth_fibonacci))

}
\seealso{
\code{\link[=get_fibonaccis]{get_fibonaccis()}}
}
