% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_polygonal_numbers.R
\name{get_polygonal_numbers}
\alias{get_polygonal_numbers}
\alias{get_triangular_numbers}
\alias{get_square_numbers}
\alias{get_pentagonal_numbers}
\alias{get_hexagonal_numbers}
\title{Provide polygonal numbers (e.g. triangular numbers)}
\usage{
get_polygonal_numbers(n, s)

get_triangular_numbers(n)

get_square_numbers(n)

get_pentagonal_numbers(n)

get_hexagonal_numbers(n)
}
\arguments{
\item{n}{the number of polygonal numbers}

\item{s}{the number of sides for the polygon}
}
\value{
a numeric vector containing the
}
\description{
Provide polygonal numbers (e.g. triangular numbers)
}
\examples{
get_polygonal_numbers(n=4, s=3)
get_triangular_numbers(n=4)
}
