% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_factors.R
\name{get_prime_factorization}
\alias{get_prime_factorization}
\title{Gets prime factorization of a number}
\usage{
get_prime_factorization(n)
}
\arguments{
\item{n}{the number to factor}
}
\value{
a list with the prime factors (\verb{$primes}) and their respective exponents (\verb{$exponents})
}
\description{
Gets prime factorization of a number
}
\examples{
pf24 = get_prime_factorization(24)
pf24$primes
pf24$exponents
prod(pf24$primes^pf24$exponents)
}
