% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_primes.R
\name{get_primes_up_to}
\alias{get_primes_up_to}
\title{Get primes numbers up to a particular number}
\usage{
get_primes_up_to(maxVal)
}
\arguments{
\item{maxVal}{the maximum values to consider}
}
\value{
a numeric vector with all primes less than or equal to \code{maxVal}
}
\description{
Get primes numbers up to a particular number
}
\examples{
get_primes_up_to(10)
}
