% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_polygonal_numbers.R
\name{is_polygonal_number}
\alias{is_polygonal_number}
\title{Tests whether a number is polygonal}
\usage{
is_polygonal_number(p, s)
}
\arguments{
\item{p}{candidate number}

\item{s}{number of polygonal sides}
}
\value{
a logical of whether p is an s-sided polygonal number
}
\description{
Tests whether a number is polygonal
}
\examples{
is_polygonal_number(p=7, s=4) # is 7 a square number?
is_polygonal_number(p=9, s=4) # is 9 a square number?
is_triangular_number <- function(p){is_polygonal_number(p, s=3)}
sapply(1:10, is_triangular_number)
}
