% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_prime.R
\name{is_prime}
\alias{is_prime}
\title{Tests to see if a numeric is prime}
\usage{
is_prime(n)
}
\arguments{
\item{n}{number to test}
}
\value{
boolean value: \code{TRUE} if \code{n} is prime and \code{FALSE} if \code{n} is composite
}
\description{
Tests to see if a numeric is prime
}
\examples{
is_prime(2)
is_prime(4)
is_prime(1)
}
