% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\name{p008_number}
\alias{p008_number}
\title{Problem 8 number}
\format{
\subsection{\code{p008_number}}{

The number (as a character) provided in Project Euler problem 8.
}
}
\source{
\url{https://projecteuler.net/problem=8}
}
\usage{
p008_number
}
\description{
The number provided in Project Euler problem 8.
}
\keyword{datasets}
