% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\name{p011_grid}
\alias{p011_grid}
\title{Problem 11 grid}
\format{
\subsection{\code{p011_grid}}{

The grid (as a numeric matrix) provided in Project Euler problem 11.
}
}
\source{
\url{https://projecteuler.net/problem=11}
}
\usage{
p011_grid
}
\description{
The grid provided in Project Euler problem 11.
}
\keyword{datasets}
