% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\name{p013_numbers}
\alias{p013_numbers}
\title{Problem 13 numbers}
\format{
\subsection{\code{p013_numbers}}{

The numbers (as a character vector) provided in Project Euler problem 13.
}
}
\source{
\url{https://projecteuler.net/problem=13}
}
\usage{
p013_numbers
}
\description{
The numbers provided in Project Euler problem 13.
}
\keyword{datasets}
