% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\name{p018_triangle}
\alias{p018_triangle}
\title{Problem 18 triangle}
\format{
\subsection{\code{p018_triangle}}{

The triangle (as a numeric matrix) provided in Project Euler problem 18.
}
}
\source{
\url{https://projecteuler.net/problem=18}
}
\usage{
p018_triangle
}
\description{
The triangle provided in Project Euler problem 18.
}
\keyword{datasets}
