% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\name{p022_names}
\alias{p022_names}
\title{Problem 22 names}
\format{
\subsection{\code{p022_names}}{

The names (as a character vector) provided in Project Euler problem 22.
}
}
\source{
\url{https://projecteuler.net/project/resources/p022_names.txt}
}
\usage{
p022_names
}
\description{
The names provided in Project Euler problem 22.
}
\keyword{datasets}
