% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\name{p042_words}
\alias{p042_words}
\title{Problem 42 words}
\format{
\subsection{\code{p042_words}}{

The words (as a character vector) provided in Project Euler problem 42.
}
}
\source{
\url{https://projecteuler.net/project/resources/p042_words.txt}
}
\usage{
p042_words
}
\description{
The words provided in Project Euler problem 42.
}
\keyword{datasets}
