% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\name{p054_pokerHands}
\alias{p054_pokerHands}
\title{Problem 54 pokerHands}
\format{
\subsection{\code{p054_pokerHands}}{

charcter matrix containing poker hands
}
}
\source{
\url{https://projecteuler.net/project/resources/p054_poker.txt}
}
\usage{
p054_pokerHands
}
\description{
The pokerHands provided in Project Euler problem 54.
}
\keyword{datasets}
