% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\name{p067_triangle}
\alias{p067_triangle}
\title{Problem 67 triangle}
\format{
\subsection{\code{p067_triangle}}{

The triangle (as a numeric matrix) provided in Project Euler problem 67.
}
}
\source{
\url{https://projecteuler.net/resources/documents/0067_triangle.txt}
}
\usage{
p067_triangle
}
\description{
The triangle provided in Project Euler problem 67.
}
\keyword{datasets}
