% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\name{p079_keylog}
\alias{p079_keylog}
\title{Problem 79 keylog}
\format{
\subsection{\code{p079_keylog}}{

The keylog (as a numeric vector) provided in Project Euler problem 79.
}
}
\source{
\url{https://projecteuler.net/resources/documents/0079_keylog.txt}
}
\usage{
p079_keylog
}
\description{
The keylog provided in Project Euler problem 79.
}
\keyword{datasets}
