% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\name{p081_matrix}
\alias{p081_matrix}
\title{Problem 81 matrix}
\format{
\subsection{\code{p081_matrix}}{

The matrix (as a numeric matrix) provided in Project Euler problem 81.
}
}
\source{
\url{https://projecteuler.net/resources/documents/0081_matrix.txt}
}
\usage{
p081_matrix
}
\description{
The matrix provided in Project Euler problem 81.
}
\keyword{datasets}
