% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\name{p082_matrix}
\alias{p082_matrix}
\title{Problem 82 matrix}
\format{
\subsection{\code{p082_matrix}}{

The matrix (as a numeric matrix) provided in Project Euler problem 82.
}
}
\source{
\url{https://projecteuler.net/resources/documents/0082_matrix.txt}
}
\usage{
p082_matrix
}
\description{
The matrix provided in Project Euler problem 82.
}
\keyword{datasets}
