% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\name{p083_matrix}
\alias{p083_matrix}
\title{Problem 83 matrix}
\format{
\subsection{\code{p083_matrix}}{

The matrix (as a numeric matrix) provided in Project Euler problem 83.
}
}
\source{
\url{https://projecteuler.net/resources/documents/0083_matrix.txt}
}
\usage{
p083_matrix
}
\description{
The matrix provided in Project Euler problem 83.
}
\keyword{datasets}
