% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\name{p089_romanNumerals}
\alias{p089_romanNumerals}
\title{Problem 89 romanNumerals}
\format{
\subsection{\code{p089_romanNumerals}}{

The roman numerals (as a character vector) provided in Project Euler problem 89.
}
}
\source{
\url{https://projecteuler.net/resources/documents/0089_roman.txt}
}
\usage{
p089_romanNumerals
}
\description{
The romanNumerals provided in Project Euler problem 89.
}
\keyword{datasets}
