% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\name{p096_sudokuGrids}
\alias{p096_sudokuGrids}
\title{Problem 96 sudokuGrids}
\format{
\subsection{\code{p096_sudokuGrids}}{

The sudoku grids (as a list of numeric matrices) provided in Project Euler problem 96.
}
}
\source{
\url{https://projecteuler.net/project/resources/p096_sudoku.txt}
}
\usage{
p096_sudokuGrids
}
\description{
The sudokuGrids provided in Project Euler problem 96.
}
\keyword{datasets}
