% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\name{p098_words}
\alias{p098_words}
\title{Problem 98 words}
\format{
\subsection{\code{p098_words}}{

The words (as a character vector) provided in Project Euler problem 98.
}
}
\source{
\url{https://projecteuler.net/resources/documents/0098_words.txt}
}
\usage{
p098_words
}
\description{
The words provided in Project Euler problem 98.
}
\keyword{datasets}
