% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{Reads Project Euler data}
\usage{
read_data(problem = NA_integer_)
}
\arguments{
\item{problem}{The problem number to load data for.}
}
\value{
Data for the problem. Varies by problem, sometimes character vectors, sometimes numeric vectors. Where it makes sense arrays are provided.
}
\description{
Reads Project Euler data
}
\examples{
read_data(problem = 22)[1:10]
}
