% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_operations.R
\name{sum_letter_values}
\alias{sum_letter_values}
\title{Provides sum of numeric substitution for letters}
\usage{
sum_letter_values(word, letterVals = stats::setNames(1:26, LETTERS))
}
\arguments{
\item{word}{the "word" whose letters to consider}

\item{letterVals}{the dictionary (a named numeric vector)}
}
\value{
the single numeric sum
}
\description{
Provides sum of numeric substitution for letters
}
\examples{
sum_letter_values("COLIN")
}
