% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_N.R
\name{sum_N}
\alias{sum_N}
\title{Adds two BIG Z numbers}
\usage{
sum_N(a, b)
}
\arguments{
\item{a}{a big Z summand}

\item{b}{a big Z summand}
}
\value{
the sum of a and b
}
\description{
Adds two BIG Z numbers
}
