% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_of_powers.R
\name{sum_of_powers}
\alias{sum_of_powers}
\title{Adds values of x to the p-th power}
\usage{
sum_of_powers(x, p)
}
\arguments{
\item{x}{values exponentiate and sum}

\item{p}{power to exponentiate with}
}
\value{
the numeric sum
}
\description{
Adds values of x to the p-th power
}
\section{May be useful for ...}{

\itemize{
\item Problem 6: Sum Square Difference
\item Problem 30: Digit fifth powers
}
}

\examples{
sum_of_powers(x=1:100, p=1) # famous sum of Gauss equal to `5050`
sum_of_powers(x=2^(1:32), p=-1) # famous sum of Zeno equal to `1`
sum_of_powers(x=1:10^6, p=-2) # famous sum of Euler equal to `pi^2/6`
}
