% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prime_factors.R
\name{get_prime_factors}
\alias{get_prime_factors}
\title{Gets the prime factors of a number.}
\usage{
get_prime_factors(n)
}
\arguments{
\item{n}{the number to factor}
}
\value{
a numeric vector containing the prime factors of n
}
\description{
Gets the prime factors of a number.
}
\examples{
get_prime_factors(7)
get_prime_factors(6)
get_prime_factors(2^10)
}
