% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_prime.R
\name{is_prime}
\alias{is_prime}
\title{Primality Testing}
\usage{
is_prime(n)
}
\arguments{
\item{n}{number to be tested}
}
\value{
returns \code{TRUE} or \code{FALSE} depending on whether its argument is prime or not
}
\description{
Primality Testing
}
\section{May be useful for}{

\itemize{
\item Problem 27: Quadratic primes
\item Problem 35: Circular primes
\item Problem 41: Pandigital prime
}
}

\examples{
is_prime(2)
is_prime(4)
is_prime(1)
is_prime(2^31 - 1)
is_prime(10^16)
}
