% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\name{p054_pokerHands}
\alias{p054_pokerHands}
\title{Problem 54 pokerHands}
\format{
\subsection{\code{p054_pokerHands}}{

A character matrix with 1,000 rows and 10 columns. The first five columns contain the cards for the first player and the last five columns contain the cards for the second player:
\describe{
\item{Pi_Cj}{the jth card for the ith player (e.g the \code{P1_C3} column contains the first player's 3rd cards)}
}
}
}
\source{
\url{https://projecteuler.net/project/resources/p054_poker.txt}
}
\usage{
p054_pokerHands
}
\description{
The pokerHands provided in Project Euler problem 54.
}
\keyword{datasets}
