% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projecteulerr-package.R
\docType{package}
\name{projecteulerr-package}
\alias{projecteulerr}
\alias{projecteulerr-package}
\title{projecteulerr: Helpful Functions for Solving Project Euler Problems}
\description{
Convenience functions for solving Project Euler problems.
}
\keyword{internal}
